+++ compiler/common/NativeLinkerFlags.ts
@@ -1,8 +1,24 @@
+export function getNativeCodegenFlags(
+  platform: NodeJS.Platform | string = process.platform,
+): string[] {
+  if (platform === "linux") {
+    return ["-ffunction-sections", "-fdata-sections"];
+  }
+
+  return [];
+}
+
 export function getNativeLinkerFlags(
   platform: NodeJS.Platform | string = process.platform,
 ): string[] {
   if (platform === "linux") {
-    return ["-lm", "-ldl", "-rdynamic"];
+    return [
+      ...getNativeCodegenFlags(platform),
+      "-Wl,--gc-sections",
+      "-Wl,--no-export-dynamic",
+      "-lm",
+      "-ldl",
+    ];
   }
 
   if (platform === "darwin") {